requirements.txt
azure-storage-blob==12.17.0
azure-cognitiveservices-vision-customvision==3.1.0
semantic-kernel==1.29.0
python-dotenv==1.0.1
pillow==10.0.0
requests==2.31.0